comments and proposed changes get saved
